refactor: move PassPlaylist into own ui folder
