+++ src/features/offerv2/components/OfferPlaylist/OfferPlaylist.tsx
@@ -4,7 +4,7 @@ import { View } from 'react-native'
 import { HitOfferWithArtistAndEan } from 'features/offer/components/OfferPlaylistOld/api/fetchOffersByArtist'
 import { PlaylistType } from 'features/offer/enums'
 import { Offer } from 'shared/offer/types'
-import { PassPlaylist } from 'ui/components/PassPlaylist'
+import { PassPlaylist } from 'ui/components/PassPlaylist/PassPlaylist'
 import { Spacer } from 'ui/theme'
 
 interface OfferPlaylistProps {
